docs: 홈 페이지 레이아웃 적용
